requirements.txt
torch
torchaudio
torchvision
torchinfo
pillow
opencv-python
matplotlib

